change text-domain, update readme
