+++ readme.txt
@@ -1,10 +1,10 @@
-=== AutomatePlus - Mautic for Give ===
+=== AutomatePlug - Mautic for Give ===
 Contributors: brainstormforce, rahulwalunje-1
 Tags: mautic, marketing, automation, donation plugin, give, email
 Donate link: https://www.brainstormforce.com/payment/
 Requires at least: 3.0
-Tested up to: 4.7.2
-Stable tag: 1.0.0
+Tested up to: 4.8
+Stable tag: 1.0.1
 License: GPLv2 or later
 License URI: https://www.gnu.org/licenses/gpl-2.0.html
 
@@ -15,9 +15,15 @@ Integrate Mautic with your Give donation forms. Add donors to Mautic segment whe
 Integrate Mautic with your Give donation forms. Add donors to Mautic segment when they donate to your Cause with Give - WordPress Donation Plugin.
 
 In order to use this plugin, First you need to install
-[AutomatePlus - Mautic for WordPress](https://wordpress.org/plugins/automate-mautic/) and [Give - WordPress Donation Plugin](https://wordpress.org/plugins/give/)
+[AutomatePlug - Mautic for WordPress](https://wordpress.org/plugins/automate-mautic/) and [Give - WordPress Donation Plugin](https://wordpress.org/plugins/give/)
 
 == Changelog ==
+
+= 1.0.1=
+
+* Optimize code and make compatible with latest AutomatePlug - Mautic for WordPress
+* Optimize subscribe mautic call
+* Translation compatible
 
 = 1.0.0 =
